my-ollama-chatbot/pyproject.toml
[tool.poetry]
name = "multi-agent-chatbot"
version = "0.1.0"
description = "Multi-agent AI chatbot with Gradio, Ollama, RAG, and image capabilities."
authors = ["Your Name <you@example.com>"]
license = "MIT"
readme = "README.md"


[tool.poetry.dependencies]
python = ">=3.10, <3.12"
gradio = "5.29.1"
langchain = "~0.1.16"
langchain-community = "~0.0.32"
langchain-core = "~0.1.42"
langgraph = "~0.0.33"
ollama = "^0.1.8"
chromadb = "^0.4.24"
pypdf = "^4.0.0"
tiktoken = "^0.6.0"
pillow = "^10.0.0"
python-dotenv = "^1.0.0" # .env 파일 관리용 (선택 사항)
unstructured = {extras = ["pdf"], version = "^0.13.0"} # PDF 내 이미지/테이블 등 복잡한 구조 처리용
pikepdf = "^8.14.0" # 가끔 pypdf가 못 여는 PDF 처리용
beautifulsoup4 = "^4.12.3" # HTML/XML 파싱 (PDF 내 JS 추출 시 필요할 수 있음)
pdfplumber = "^0.11.6"
pydantic = ">=2.0,<2.12"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
ruff = "^0.9" # Python Linter

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"